feat(core): Implemented TagFactory and refactored NodeDeviceDTO operationMode and reportingPeriod Modules.
- Added two different configurations dev and prod to use with different repository implementations.

+++ src/main/java/com/acousea/backend/app/api/v1/communicationSystem/CommunicationWebhookController.java
@@ -38,7 +38,7 @@ public class CommunicationWebhookController {
             @RequestParam("transmit_time") String transmitTime,
             @RequestParam("iridium_latitude") double iridiumLatitude,
             @RequestParam("iridium_longitude") double iridiumLongitude,
-            @RequestParam("iridium_cep") int iridiumCep,
+            @RequestParam("iridium_cep") double iridiumCep,
             @RequestParam("data") String data) {
         try {
             // Crear un objeto RockBlockMessage con los parámetros recibidos
